requirements.txt
streamlit>=1.30.0
matplotlib==3.8.3
pandas==2.2.0
seaborn==0.13.1
numpy>=1.24.0
python-dotenv>=1.0.0
PyYAML>=6.0
pytz==2023.3
fpdf==1.7.2
openai==0.28.0 